+++ src/main/java/com/everestengg/code/challenge/client/PackageDeliveryCostEstimationApp.java
@@ -1,62 +1,52 @@
 package com.everestengg.code.challenge.client;
 
-import com.everestengg.code.challenge.util.IoUtils;
-import com.everestengg.code.challenge.vo.PackageDeliveryInput;
+import com.everestengg.code.challenge.model.PackageDeliveryCostEstimateInfo;
+import com.everestengg.code.challenge.repo.StaticOfferRepository;
+import com.everestengg.code.challenge.service.delivery.cost.estimation.PackageDeliveryCostEstimationServiceFactory;
+import com.everestengg.code.challenge.io.IoUtils;
 import com.everestengg.code.challenge.vo.InputPackage;
-import com.everestengg.code.challenge.model.PackageDeliveryTimeEstimationInfo;
-import com.everestengg.code.challenge.service.delivery.time.estimation.PackageDeliveryTimeEstimationServiceFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
+import java.util.Scanner;
 
-
-import static com.everestengg.code.challenge.repo.StaticOfferRepository.prepareOffers;
-import static com.everestengg.code.challenge.service.delivery.time.estimation.PackageDeliveryTimeEstimationServiceFactory.PackageDeliveryTimeEstimationType.SIMPLE;
-
+import static com.everestengg.code.challenge.service.delivery.cost.estimation.PackageDeliveryCostEstimationServiceFactory.PackageServiceType.SIMPLE;
 
 public class PackageDeliveryCostEstimationApp {
     private static final Logger logger = LoggerFactory.getLogger(PackageDeliveryCostEstimationApp.class);
     public static void main(String[] args) {
         prepareOffers();
         try(Scanner scanner = new Scanner(System.in)) {
-            /**
-             * Read base delivery cost and no of packages
-             */
-            System.out.println("enter base delivery cost and no of packages");
-            short baseDeliveryCost = scanner.nextShort();
-            short noOfPackages = scanner.nextShort();
+            short baseDeliveryCost = IoUtils.readBaseCost(scanner);
+            short noOfPackages = IoUtils.readNoOfPackages(scanner);
             InputPackage[] inputPackages = new InputPackage[noOfPackages];
             // Read packages
-            readPackages(scanner, noOfPackages, inputPackages);
-
-            // read delivery information
-            PackageDeliveryInput packageDeliveryInput = readDeliveryInput(scanner);
+            IoUtils.readPackages(scanner, noOfPackages, inputPackages);
             //calculate discount for each and print
-            List<PackageDeliveryTimeEstimationInfo> result = PackageDeliveryTimeEstimationServiceFactory
-                    .getDeliveryEstimationService(SIMPLE).calculateEstimatedDelivery(inputPackages, packageDeliveryInput,
-                            baseDeliveryCost);
-            printDelivery(result);
+            process(baseDeliveryCost, inputPackages);
         }catch (Exception e){
             logger.error("app crashed.. error ",e);
         }
 
     }
 
-    private static void readPackages(Scanner scanner, short noOfPackages, InputPackage[] inputPackages) {
-        IoUtils.readPackages(scanner, noOfPackages, inputPackages);
-    }
-    private static void printDelivery(List<PackageDeliveryTimeEstimationInfo> packageDeliveryTimeEstimationInfo) {
-        for (PackageDeliveryTimeEstimationInfo packageDeliveryTimeEstimationInfo1 : packageDeliveryTimeEstimationInfo) {
-            System.out.println(packageDeliveryTimeEstimationInfo1);
+
+
+
+
+    private static void process(short baseDeliveryCost, InputPackage[] inputPackages) {
+        for (InputPackage inputPackage : inputPackages){
+            PackageDeliveryCostEstimateInfo packageDeliveryCostEstimateInfo = PackageDeliveryCostEstimationServiceFactory
+                    .getPackageOrderService(SIMPLE).calcCost(inputPackage, baseDeliveryCost).getResult();
+            packageDeliveryCostEstimateInfo.print();
         }
     }
 
 
-    private static PackageDeliveryInput readDeliveryInput(Scanner scanner) {
-        System.out.println("enter no of vehicles, max speed and max carriable weight");
-        return PackageDeliveryInput.builder().noOfVehicle(scanner.nextShort()).maxSpeed(scanner.nextShort())
-                .maxCarriableWt(scanner.nextShort()).build();
+    /**
+     * Offer preparation below
+     */
+    private static void prepareOffers() {
+      StaticOfferRepository.prepareOffers();
     }
 }
-
