+++ src/main/java/com/everestengg/code/challenge/vo/PackageDeliveryInput.java
@@ -13,4 +13,14 @@ public class PackageDeliveryInput {
     private short noOfVehicle;
     private short maxSpeed;
     private short maxCarriableWt;
+
+    public boolean isNoOfVehicleValid(){
+        return noOfVehicle > 0;
+    }
+    public boolean isValidMaxSpeed(){
+        return maxSpeed > 0;
+    }
+    public boolean isValidMaxCarriableWt(){
+        return maxCarriableWt > 0;
+    }
 }
